+++ lib/features/children/add_member/cubit/add_member_cubit.dart
@@ -14,36 +14,68 @@ class AddMemberCubit extends Cubit<AddMemberState> {
 
 // to do add many children
   void rememberChild({required Child child}) {
+    final existingChildIndex =
+        state.children.indexWhere((c) => c.key == child.key);
+
+    if (existingChildIndex != -1) {
+      // Child with the same key exists, replace it
+      final updatedChildren = state.children;
+      updatedChildren[existingChildIndex] = child;
+
+      emit(
+        state.copyWith(
+          children: updatedChildren,
+          status: AddMemberStateStatus.input,
+        ),
+      );
+    } else {
+      // Child with the key doesn't exist, add it
+      emit(
+        state.copyWith(
+          children: state.children..add(child),
+          status: AddMemberStateStatus.input,
+        ),
+      );
+    }
+  }
+
+  void validateForms() {
     emit(
       state.copyWith(
-        child: child,
         status: AddMemberStateStatus.input,
+        formStatus: AddMemberFormStatus.validate,
       ),
     );
   }
 
-  void goToVPC(Child? child) {
-    emit(state.copyWith(
-      status: AddMemberStateStatus.vpc,
-      child: child ?? state.child,
-    ));
+  void resetFormStatus() {
+    emit(
+      state.copyWith(
+        status: AddMemberStateStatus.input,
+        formStatus: AddMemberFormStatus.initial,
+      ),
+    );
   }
 
+// TODO: display children in the forms when navigating back from VPC
+// or navigate all the way back to the family overview page
   void goToInput() {
     emit(state.copyWith(
       status: AddMemberStateStatus.input,
+      formStatus: AddMemberFormStatus.initial,
+      children: [],
     ));
   }
 
   Future<void> createChildWithVPC() async {
-    final child = state.child;
+    final children = state.children;
     emit(
       state.copyWith(
         status: AddMemberStateStatus.loading,
       ),
     );
     try {
-      final isChildCreated = await _addMemberRepository.createChild(child);
+      final isChildCreated = await _addMemberRepository.addMembers(children);
       if (isChildCreated) {
         emit(
           state.copyWith(
