+++ lib/features/children/add_member/repository/add_member_repository.dart
@@ -2,7 +2,7 @@ import 'package:givt_app/core/network/api_service.dart';
 import 'package:givt_app/features/children/add_member/models/child.dart';
 
 mixin AddMemberRepository {
-  Future<bool> createChild(Child child);
+  Future<bool> addMembers(List<Child> children);
 }
 
 class AddMemberRepositoryImpl with AddMemberRepository {
@@ -10,12 +10,13 @@ class AddMemberRepositoryImpl with AddMemberRepository {
   final APIService apiService;
 
   @override
-  Future<bool> createChild(Child child) async {
+  Future<bool> addMembers(List<Child> children) async {
+    final List<Map<String, dynamic>> profilesJsonList =
+        children.map((child) => child.toJson()).toList();
     final body = {
-      'profiles': [
-        child.toJson(),
-      ]
+      'profiles': profilesJsonList,
     };
+
     final response = await apiService.addMember(body);
     return response;
   }
