+++ lib/features/children/overview/widgets/no_children_page.dart
@@ -39,7 +39,7 @@ class NoChildrenPage extends StatelessWidget {
         ),
         const SizedBox(height: 8),
         Padding(
-          padding: const EdgeInsets.only(top: 20, left: 35, right: 35),
+          padding: const EdgeInsets.only(top: 20, left: 35, right: 35, bottom: 12),
           child: ElevatedButton(
             onPressed: onAddNewChildPressed,
             child: Text(
